+++ src/rules/rules2ldl_main.ml
@@ -14,9 +14,10 @@
  * limitations under the License.
  *)
 
-module Ldl = Ldlsat.Ldl
-
+open Dsl4sc
 open Printf
+
+module Ldl = Ldlsat.Ldl
 
 let stdin = open_in "/dev/stdin"
 let opt_o = ref "unspecified"