+++ tools/modelgen/src/modelgen_main.ml
@@ -13,6 +13,7 @@
  * limitations under the License.
  *)
 
+open Dsl4sc
 open Printf
 
 let stdin = open_in "/dev/stdin"